Create a new ModifiedBrightnessIndicator every time.
Because we force calling run_dispose() when swapping out a brightness
indicator (to fully disconnect the signal), we cannot just use a
half-destroyed object.
Split out the code overriding the brightness indicator methods in its
own class, and instantiate that ModifiedBrightnessIndicator every time
the extension is enabled.
Fixes the bug where the extension would fail if initially enabled,
then disabled and reenabled.