Tentative bug-fix for the extension stopping working
The extension sometimes stops working after a resume on laptops.
Disabling and reenabling the extension does not fix it, but restarting
the shell does.  It seems that the background color of the label
providing the overlay gets lost.  Since it's obtained by CSS, remove
stylesheet.css and hard-code the CSS properties directly when creating
the labels.