Brightness applies to the cursor / mouse pointer.
The real cursor is actually hidden, and we show a clone of it behind the
overlay layer.
This technique comes from gnome-shell's Magnifier implementation.
Monkey patch CursorTracker to prevent the real cursor from showing up.

There is still a problem where seemingly C code unconditionally turns
the cursor back on in a way that cannot be detected by the extension.
Work around that by re-hiding the cursor on every mouse motion and cursor
change: _updateMousePosition, _updateMouseSprite and
_cursorTrackerSetPointerVisibleReplacement call _setPointerVisible(false)
which is probably not the best for performance.