Monitor stage actors and simplify magnifier handling
Just monitor the stage actors being added and removed and raise
ourselves on top on them whenever this happens.
We can get rid of all the patching for the magnifier since the
overlay will now be stacked on top of it and will affect it without
any further intervention needed.