Attempt to fix cursor flashing bug #2 by delaying hiding the cursor.
Whenever the cursor gets hidden by mouse motion or cursor change,
not only immediately hide the cursor, but also again at the beginning
of the next redraw.
Now while the flashing is almost gone, it still happens every one in
a while especially when moving around title bars and switching windows
with alt-tab.