Add support for gnome-shell 43
gnome-shell 43 replaces AggregateMenu with QuickSettings:
https://gjs.guide/extensions/upgrading/gnome-shell-43.html#quick-settings

The current strategy used for modifying brightness indicator
behaviour is as follows.  When the extension is enabled, the
existing brightness indicator is destroyed, and a new one
extending imports.ui.status.brightness.Indicator is injected into
AggregateMenu, replacing the original one.  When the extension is
disabled, an "official" brightness indicator object is
re-initialized, and injected back to its original position in
AggregateMenu.

When considering how to implement GNOME 43 support, several
considerations went into the decision of moving from the current
approach to a "monkey-patching" approach.

First of all, the brightness indicator module moves the bulk of
its functionality to a private class:
https://gitlab.gnome.org/GNOME/gnome-shell/-/commit/7bbd59838a6cfe14b189fa7bd8e743fb0cac9bc3

Strictly speaking, the class is still accessible and could be
extended in a way similar to current code, but accessing the
class causes a warning to be printed to the console: "The
property access will work as previously for the time being, but
please fix your code anyway."

The brightness indicator code was also changed quite
significantly, and a wrapper class is added in gnome-shell 43.
That suggests two possible approaches:
  (1) Keep extending for GS 42-, but copy-and-modify for GS 43+
  (2) Copy-and-modify for both, with conditional clauses for
      42-/43+ differences
Both of these approaches result in quite a lot of gnome-shell
code being duplicated.  Copying GS 43+ code could also be quite
fragile if future releases make any changes to the brightness
indicator.

As such, a "monkey-patching" approach is used in this commit,
minimizing the differences between GS 42-/43+ clauses.  The
biggest downside to this approach is dealing with some weirdness
surrounding the Brightness proxy object and how it calls the
_sync function (see comments in code for a full explanation).