Change ScreenshotClass to access softBrightnessExtension via closure
Wrap ScreenshotClass definition in a function which accepts
softBrightnessExtension as an argument.  This allows the caller to pass
a reference to the extension class without exposing it as a global.

This may not be an ideal solution since it calls GObject.registerClass
every time the extension is enabled.  On the other hand, I could not
find instances of other extensions "unregistering" classes.