Revert "Change ScreenshotClass to access softBrightnessExtension via closure"
This reverts commit e1af5ddf86ca59706b3ee6c5827cb2c7b063419a.

Wrapping ScreenshotClass's definition in a closure is causing issues
when disabling/re-enabling the extension.  When re-enabling,
GObject.registerClass gets called on ScreenshotClass for a second
time, which throws an error.

AFAIK there is no way of "unregistering" a class with GObject.  Thus
classes cannot be registered as part of an extension's enable()
function.  Other extensions seem to register the class as part of the
extension's module code, which only runs once when gnome-shell loads.

Possible options:
(A) Switch to a delegate pattern [1]
(B) Subclass Shell.Screenshot, but keep definition in module namespace

Since official documentation talks about migrating from native
JavaScript classes (such as those used for a delegate pattern) to
GObject classes [2], choose option (B).

[1] https://bugzilla.gnome.org/show_bug.cgi?id=688973
[2] https://wiki.gnome.org/Projects/GnomeShell/Extensions/MigratingShellClasses