Always follow gnome-shell enable/disable calls
Previously, in order to support the extension running on the lock
screen, disable() was ignored in GNOME 41 and below.

In order to simplify code paths, and to adhere to the GNOME extension
guidelines[1], this functionality is being removed.  The extension will
no longer run on the lock screen in GNOME 41 and below.

[1] https://gjs.guide/extensions/review-guidelines/review-guidelines.html#session-modes