Specify width and height of preferences window
After updating to extend from ExtensionPreferences, the default window
size is not large enough to fit the preferences Gtk.Grid, and Adw
repeatedly complains, spitting out error messages whenever the window is
resized.

I am not a GTK expert, and I believe there must be a better way of doing
this than hard-coding a minimum width and height, but this seems to be
what all the other extensions are doing.  So for now, we shall set the
minimum width and height in `fillPreferencesWindow`.

Also take `text-scaling-factor` into account.