Enable: wait for quickSettings._brightness to become available
Previously when starting up gnome-shell, quickSettings._brightness
was set synchronously, and thus it was guaranteed to exist by the time
extensions started getting enabled.

Now, all of the quickSettings._* properties are set asynchronously:
https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2563

enable() can either:
(1) Use the same method that is used for cursor cloning, and set a
    timeout which runs code later on (when _brightness is available)
    to finish initialization.
(2) Itself become an asynchronous function, and wait for _brightness
    to become available before returning.

(1) has the downside that it becomes difficult to reason about what
state the extension is in (e.g. when disable is called), whereas (2)
has the benefit that after enable() completes, we are guaranteed that
the extension has really completed enabling.

Convert enable() to be an async function, and use a simple
await Promise(setTimeout) construct to implement a wait-retry loop.

disable() code can be simplified, since it is guaranteed that cursor
cloning initialization has already completed.